Added option to block play/pause click on video
Added an option that puts a div over the iframe to effectively block
clicks on the player itself. This does prevent the right click too
however. I needed to block the ability to play/pause with the left click
because I wanted to track when user does those actions via the control
bar. Don't use this options with ytcontrols enabled.