Don't destroy the YouTube player on error (#400)
Features were added to videojs-youtube which allows videos to be loaded once the player is instantiated, however the player is destroyed whenever a player error occurs.  This means that calling player.src('') with some new video will no longer work when an error has occurred on the current video.  This PR removes the destruction of the player when an error occurs, and also destroys it properly whenever dispose() is called on the tech.
* Don't destroy the YouTube player on error
* Call Tech.prototype.dispose() on the YouTube tech when disposing