Fix massive performance issue with large videos
Time range length property (returned by`buffered` and `seekable`)
should represent the number of time ranges our start and end
function will respond to[1]. In this case, we only know the
percentage loaded so we're assuming that it starts at the beginning.

I think that logic is flawed but that may be the best possible solution
given what the YouTube API returns.

In any event, using the duration as the `length` makes VideoJS loop
`length` times to call `start` and `end`. This happens a few times
per second, making the video barely usable for very long videos.

[1] If I understand correctly.